# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP-сервер на Go с автоматической генерацией Swagger при старте",
        "title": "Configurator",
        "contact": {
            "name": "Lorgar Avrelian",
            "url": "https://github.com/Lorgar-Avrelian",
            "email": "victor-14-244@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://apache.org"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "paths": {
...
        "/api/v1/params": {
            "post": {
                "description": "Создает новый системный параметр в таблице public.param",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "2. Модельный каталог: Параметры"
                ],
                "summary": "Создать параметр",
                "parameters": [
                    {
                        "description": "Данные параметра",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/configurator_internal_dto.ParamCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/configurator_internal_model.Param"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка базы данных",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/params/search": {
            "get": {
                "description": "Ищет параметры, у которых title, name_en, name_ru, description_en или description_ru частично совпадают с переданной строкой query",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "2. Модельный каталог: Параметры"
                ],
                "summary": "Поиск параметров по строке",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Строка поиска",
                        "name": "query",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/configurator_internal_model.Param"
                            }
                        }
                    },
                    "400": {
                        "description": "Пустой запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка базы данных",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/params/unattached": {
            "get": {
                "description": "Возвращает список всех параметров, которые не привязаны ни к одной составной части устройства",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "2. Модельный каталог: Параметры"
                ],
                "summary": "Получить непривязанные параметры",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/configurator_internal_model.Param"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка базы данных",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/params/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "2. Модельный каталог: Параметры"
                ],
                "summary": "Получить параметр по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Параметра",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/configurator_internal_model.Param"
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Параметр не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка базы данных",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "2. Модельный каталог: Параметры"
                ],
                "summary": "Обновить параметр",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Параметра",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/configurator_internal_dto.ParamUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/configurator_internal_model.Param"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации или неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Параметр не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка базы данных",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "2. Модельный каталог: Параметры"
                ],
                "summary": "Удалить параметр",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Параметра",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Параметр не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка базы данных",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "configurator_internal_dto.ParamCreate": {
            "type": "object",
            "required": [
                "access",
                "name_en",
                "name_ru",
                "title",
                "type"
            ],
            "properties": {
                "access": {
                    "type": "string",
                    "example": "USER"
                },
                "description_en": {
                    "type": "string",
                    "example": "Operational Voltage"
                },
                "description_ru": {
                    "type": "string",
                    "example": "Рабочее напряжение"
                },
                "name_en": {
                    "type": "string",
                    "example": "Voltage"
                },
                "name_ru": {
                    "type": "string",
                    "example": "Напряжение"
                },
                "saved": {
                    "type": "boolean",
                    "example": true
                },
                "title": {
                    "type": "string",
                    "example": "voltage"
                },
                "type": {
                    "type": "string",
                    "example": "NUMERIC"
                },
                "units_en": {
                    "type": "string",
                    "example": "V"
                },
                "units_ru": {
                    "type": "string",
                    "example": "В"
                },
                "value": {
                    "type": "string",
                    "example": "220"
                },
                "visible": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "configurator_internal_dto.ParamUpdate": {
            "type": "object",
            "required": [
                "access",
                "name_en",
                "name_ru",
                "title",
                "type"
            ],
            "properties": {
                "access": {
                    "type": "string",
                    "example": "ADMIN"
                },
                "description_en": {
                    "type": "string",
                    "example": "Maximum operational voltage"
                },
                "description_ru": {
                    "type": "string",
                    "example": "Максимальное рабочее напряжение"
                },
                "name_en": {
                    "type": "string",
                    "example": "max_voltage"
                },
                "name_ru": {
                    "type": "string",
                    "example": "макс_напряжение"
                },
                "saved": {
                    "type": "boolean",
                    "example": true
                },
                "title": {
                    "type": "string",
                    "example": "Макс. Напряжение"
                },
                "type": {
                    "type": "string",
                    "example": "INT"
                },
                "units_en": {
                    "type": "string",
                    "example": "V"
                },
                "units_ru": {
                    "type": "string",
                    "example": "В"
                },
                "value": {
                    "type": "string",
                    "example": "240"
                },
                "visible": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "configurator_internal_model.Param": {
            "type": "object",
            "properties": {
                "access": {
                    "type": "string",
                    "example": "ADMIN"
                },
                "description_en": {
                    "type": "string",
                    "example": "Maximum operational voltage"
                },
                "description_ru": {
                    "type": "string",
                    "example": "Максимальное рабочее напряжение"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name_en": {
                    "type": "string",
                    "example": "max_voltage"
                },
                "name_ru": {
                    "type": "string",
                    "example": "макс_напряжение"
                },
                "saved": {
                    "type": "boolean",
                    "example": true
                },
                "title": {
                    "type": "string",
                    "example": "Макс. Напряжение"
                },
                "type": {
                    "type": "string",
                    "example": "INT"
                },
                "units_en": {
                    "type": "string",
                    "example": "V"
                },
                "units_ru": {
                    "type": "string",
                    "example": "В"
                },
                "value": {
                    "type": "string",
                    "example": "220"
                },
                "visible": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
...
}